Move the `Rotational` component to a separate subfolder
